fixed place for drawing and show action for large objects

+++ Engine/DayNightCycle.cs
@@ -15,7 +15,7 @@ namespace Engine
         public static int DaysInMonth = 10;
         public static int MonthesInYear = 12;
         public static int TicksInSecond = Game.TimeStep/10 == 0 ? 1 : Game.TimeStep / 10;
-        public static int DayLength = TicksInSecond * (int)MinutesInHour * HoursInDay;
+        public static int DayLength = TicksInSecond * MinutesInHour * HoursInDay;
         public static int HalfDayLength = DayLength/2;
         public static int OneSixDayLength = DayLength/6;
         public static int OneSevenDayLength = DayLength/7;
